requirements.txt
fastapi
uvicorn
razorpay
# sqlite3
python-dotenv
pydantic